organize a bit tests
