+++ src/main/scala/mg7/bundles.scala
@@ -2,7 +2,6 @@ package ohnosequences.mg7
 
 import ohnosequences.statika._
 
-
 case object bundles {
 
   case object flash extends ohnosequencesBundles.statika.Flash("1.2.11")
