+++ src/test/scala/mg7/PRJEB6592/PRJEB6592.scala
@@ -1,13 +1,12 @@
 package ohnosequences.test.mg7
 
+import ohnosequences.test.mg7.testDefaults._
 import ohnosequences.mg7._, loquats._, dataflows._
 import ohnosequences.datasets._, illumina._
 import ohnosequences.cosas._, types._, klists._
 import ohnosequences.loquat._
 import ohnosequences.statika._, aws._
 import ohnosequences.blast.api._
-import ohnosequences.db.rna16s
-
 import ohnosequences.awstools.ec2._, InstanceType._
 import ohnosequences.awstools.s3._
 import ohnosequences.awstools.autoscaling._
@@ -15,6 +14,24 @@ import ohnosequences.awstools.regions.Region._
 import com.amazonaws.auth._, profile._
 
 case object PRJEB6592 {
+
+  val commonS3Prefix = S3Folder("resources.ohnosequences.com", "16s/public-datasets/PRJEB6592")
+
+  val sampleIds: List[ID] = List("ERR567374")
+
+  def testOutS3Folder(sampleId: SampleID, step: StepName): S3Folder =
+    commonS3Prefix / s"${step}-test" / sampleId /
+
+  lazy val inputSamples: Map[ID, (S3Resource, S3Resource)] = sampleIds.map { id =>
+    id -> ((
+      S3Resource(commonS3Prefix / "reads" / s"${id}_1.fastq.gz"),
+      S3Resource(commonS3Prefix / "reads" / s"${id}_2.fastq.gz")
+    ))
+  }.toMap
+
+  lazy val splitInputs: Map[ID, S3Resource] = sampleIds.map { sampleId =>
+    sampleId -> S3Resource(commonS3Prefix / "flash-test" / s"${sampleId}.merged.fastq")
+  }.toMap
 
   case object testParameters extends MG7Parameters(
     outputS3Folder  = testOutS3Folder,
@@ -32,64 +49,6 @@ case object PRJEB6592 {
       parseInt(row.select(outputFields.bitscore)).map{ _ > 42 }.getOrElse(false)
     }
   }
-
-  val defaultAMI = AmazonLinuxAMI(Ireland, HVM, InstanceStore)
-
-  trait AnyTestLoquatConfig extends AnyLoquatConfig { config =>
-
-    val metadata: AnyArtifactMetadata = generated.metadata.mg7
-
-    val iamRoleName = "loquat.testing"
-    val logsBucketName = "loquat.testing"
-
-    val  managerConfig = ManagerConfig(
-      InstanceSpecs(defaultAMI, m3.medium),
-      purchaseModel = Spot(maxPrice = Some(0.1))
-    )
-
-    val workersConfig: AnyWorkersConfig = WorkersConfig(
-      instanceSpecs = InstanceSpecs(defaultAMI, m3.medium),
-      purchaseModel = Spot(maxPrice = Some(0.1)),
-      groupSize = AutoScalingGroupSize(0, 1, 10)
-    )
-
-    val terminationConfig = TerminationConfig(
-      terminateAfterInitialDataMappings = true
-    )
-
-    val dataMappings: List[AnyDataMapping]
-  }
-
-  abstract class TestLoquatConfig(
-    val loquatName: String,
-    val dataMappings: List[AnyDataMapping]
-  ) extends AnyTestLoquatConfig
-
-
-  val loquatUser = LoquatUser(
-    email = "aalekhin@ohnosequences.com",
-    localCredentials = new ProfileCredentialsProvider("default"),
-    keypairName = "aalekhin"
-  )
-
-
-  val commonS3Prefix = S3Folder("resources.ohnosequences.com", "16s/public-datasets/PRJEB6592")
-
-  val sampleIds: List[ID] = List("ERR567374")
-
-  def testOutS3Folder(sampleId: SampleID, step: StepName): S3Folder =
-    commonS3Prefix / s"${step}-test" / sampleId /
-
-  val inputSamples: Map[ID, (S3Resource, S3Resource)] = sampleIds.map { id =>
-    id -> ((
-      S3Resource(commonS3Prefix / "reads" / s"${id}_1.fastq.gz"),
-      S3Resource(commonS3Prefix / "reads" / s"${id}_2.fastq.gz")
-    ))
-  }.toMap
-
-  val splitInputs: Map[ID, S3Resource] = sampleIds.map { sampleId =>
-    sampleId -> S3Resource(commonS3Prefix / "flash-test" / s"${sampleId}.merged.fastq")
-  }.toMap
 
   val dataflow = NoFlashDataflow(testParameters)(splitInputs)
 
@@ -113,7 +72,7 @@ case object PRJEB6592 {
 
     override lazy val amiEnv = amznAMIEnv(ami, javaHeap = 3)
 
-    override val workersConfig: AnyWorkersConfig = WorkersConfig(
+    override lazy val workersConfig: AnyWorkersConfig = WorkersConfig(
       instanceSpecs = InstanceSpecs(defaultAMI, r3.large),
       purchaseModel = Spot(maxPrice = Some(0.4)),
       groupSize = AutoScalingGroupSize(0, 1, 10)
